requirements.txt
streamlit
pandas
groq
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
matplotlib